Add version timestamp to force cache invalidation
Added meta version tag with timestamp to force browsers and CDN to treat
each deployment as a completely new file. This should prevent the browser
from serving stale cached versions of the JavaScript.

Also added version log message in renderizarEscala() so you can see in the
console which version is actually running.